updated from int to double

+++ Auckland/Day 2/Weather Bot/Completed project/Weather Bot/Models/WeatherObject.cs
@@ -24,16 +24,16 @@ namespace Weather_Bot.Models
         public class Main
         {
             public double temp { get; set; }
-            public int pressure { get; set; }
-            public int humidity { get; set; }
-            public int temp_min { get; set; }
+            public double pressure { get; set; }
+            public double humidity { get; set; }
+            public double temp_min { get; set; }
             public double temp_max { get; set; }
         }
 
         public class Wind
         {
             public double speed { get; set; }
-            public int deg { get; set; }
+            public double deg { get; set; }
         }
 
         public class Clouds
